rename project files to big_int instead of std-big-int
