+++ CMakeLists.txt
@@ -3,8 +3,7 @@
 cmake_minimum_required(VERSION 3.30...4.3)
 
 project(
-    beman.std-big-int # CMake Project Name, which is also the name of the top-level
-    # targets (e.g., library, executable, etc.).
+    beman.big_int
     DESCRIPTION "Reference implementation of std::big_int"
     LANGUAGES CXX
     VERSION 0.1.0
@@ -22,14 +21,14 @@ FetchContent_MakeAvailable(googletest)
 
 # [CMAKE.SKIP_TESTS]
 option(
-    BEMAN_STD-BIG-INT_BUILD_TESTS
+    BEMAN_BIG_INT_BUILD_TESTS
     "Enable building tests and test infrastructure. Default: ${PROJECT_IS_TOP_LEVEL}. Values: { ON, OFF }."
     ${PROJECT_IS_TOP_LEVEL}
 )
 
 # [CMAKE.SKIP_EXAMPLES]
 option(
-    BEMAN_STD-BIG-INT_BUILD_EXAMPLES
+    BEMAN_BIG_INT_BUILD_EXAMPLES
     "Enable building examples. Default: ${PROJECT_IS_TOP_LEVEL}. Values: { ON, OFF }."
     ${PROJECT_IS_TOP_LEVEL}
 )
@@ -37,28 +36,28 @@ option(
 # for find of beman-install-library
 include(infra/cmake/beman-install-library.cmake)
 
-add_library(beman.std-big-int INTERFACE)
-add_library(beman::std-big-int ALIAS beman.std-big-int)
+add_library(beman.big_int INTERFACE)
+add_library(beman::big_int ALIAS beman.big_int)
 
 target_sources(
-    beman.std-big-int
+    beman.big_int
     PUBLIC FILE_SET HEADERS BASE_DIRS "${CMAKE_CURRENT_SOURCE_DIR}/include"
 )
 
 set_target_properties(
-    beman.std-big-int
+    beman.big_int
     PROPERTIES VERIFY_INTERFACE_HEADER_SETS ${PROJECT_IS_TOP_LEVEL}
 )
 
-add_subdirectory(include/beman/std-big-int)
+add_subdirectory(include/beman/big_int)
 
-beman_install_library(beman.std-big-int TARGETS beman.std-big-int)
+beman_install_library(beman.big_int TARGETS beman.big_int)
 
-if(BEMAN_STD-BIG-INT_BUILD_TESTS)
+if(BEMAN_BIG_INT_BUILD_TESTS)
     enable_testing()
-    add_subdirectory(tests/beman/std-big-int)
+    add_subdirectory(tests/beman/big_int)
 endif()
 
-if(BEMAN_STD-BIG-INT_BUILD_EXAMPLES)
+if(BEMAN_BIG_INT_BUILD_EXAMPLES)
     add_subdirectory(examples)
 endif()
