+++ examples/todo.cpp
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include <beman/std-big-int/todo.hpp>
+#include <beman/big_int/todo.hpp>
 
 int main() {
     // TODO
