+++ include/beman/big_int/CMakeLists.txt
@@ -1,0 +1,3 @@
+# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+
+target_sources(beman.big_int PUBLIC FILE_SET HEADERS FILES big_int.hpp todo.hpp)
