--- include/beman/std-big-int/CMakeLists.txt
@@ -1,6 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-
-target_sources(
-    beman.std-big-int
-    PUBLIC FILE_SET HEADERS FILES std-big-int.hpp todo.hpp
-)
